% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formats.R
\name{letter}
\alias{letter}
\alias{exam}
\alias{paper}
\alias{report}
\alias{asce}
\alias{syllabus}
\alias{slides}
\title{Output document formats for ASU Bio-inspired Geotechnics group}
\usage{
letter(...)

exam(...)

paper(...)

report(...)

asce(...)

syllabus(...)

slides(...)
}
\arguments{
\item{\dots}{Arguments passed to \code{\link[bookdown]{pdf_document2}}.}
}
\value{
An R Markdown output format object.
}
\description{
Each function is a wrapper for \code{\link[bookdown]{pdf_document2}} to
produce documents ASU BiG style.
}
\author{
Julian Tao
}
